Generate parameters

Function to generate the global parameters used in the TROLL simulation. All parameters have a default value used in French Guiana simulations.

generate_parameters(
  cols = 200,
  rows = 200,
  HEIGHT = 70,
  length_dcell = 25,
  nbiter,
  iterperyear = 12,
  NV = 1,
  NH = 1,
  nbout = 4,
  nbspp = 45,
  SWtoPPFD = 2.27,
  p_nonvert = 0.05,
  klight = 0.63,
  phi = 0.093,
  absorptance_leaves = 0.9,
  theta = 0.7,
  g1 = 3.77,
  vC = 0.021,
  DBH0 = 0.005,
  H0 = 0.95,
  CR_min = 0.3,
  CR_a = 2.13,
  CR_b = 0.63,
  CD_a = 0,
  CD_b = 0.2,
  CD0 = 0.3,
  shape_crown = 0.72,
  dens = 1,
  fallocwood = 0.35,
  falloccanopy = 0.25,
  Cseedrain = 50000,
  nbs0 = 10,
  sigma_height = 0,
  sigma_CR = 0,
  sigma_CD = 0,
  sigma_P = 0,
  sigma_N = 0,
  sigma_LMA = 0,
  sigma_wsg = 0,
  sigma_dbhmax = 0,
  corr_CR_height = 0,
  corr_N_P = 0,
  corr_N_LMA = 0,
  corr_P_LMA = 0,
  leafdem_resolution = 30,
  p_tfsecondary = 1,
  hurt_decay = 0,
  crown_gap_fraction = 0.15,
  m = 0.013,
  m1 = 0.013,
  Cair = 400,
  LL_parameterization = 1,
  LA_regulation = 2,
  sapwood = 1,
  seedsadditional = 0,
  NONRANDOM = 1,
  GPPcrown = 0,
  BASICTREEFALL = 1,
  SEEDTRADEOFF = 0,
  CROWN_MM = 0,
  OUTPUT_extended = 1,
  extent_visual = 0
)

Arguments

cols

num. Number of columns.

rows

num. Number of rows.

HEIGHT

num. Vertical extent of simulation.

length_dcell

num. Linear size of a dcell.

nbiter

num. Total number of timesteps.

iterperyear

num. Number of iterations per year.

NV

num. Vertical number of cells (per m).

NH

num. Horizontal number of cells (per m).

nbout

num. Number of outputs.

nbspp

num. Number of species

SWtoPPFD

num. Convert shortwave irradiance to PAR photons.

p_nonvert

num. Light incidence parameter (difference through turbid medium).

klight

num. Light attenuation in the canopy following a Beer-Lambert law.

phi

num. Quantum yield (in micromol C/micromol photon).

absorptance_leaves

num. Absorptance of individual leaves.

theta

num. Parameter of the Farquhar model.

g1

num. Parameter g1 of Medlyn et al stomatal conductance model.

vC

num. Variance of the flexion moment.

DBH0

num. Initial diameter at breast height (m).

H0

num. Initial height (m).

CR_min

num. Minimum crown radius (in m).

CR_a

num. Crown radius log intercept or Michaelis Menten initial growth.

CR_b

num. Crown radius log slope or Michaelis Menten asymptotic CR.

CD_a

num. Crown depth intercept (absolute value).

CD_b

num. Crown depth slope (as fraction of tree height).

CD0

num. Initial crown depth (in m).

shape_crown

num. Crown shape parameter.

dens

num. Initial leaf density (m^2/m^2).

fallocwood

num. Fraction of biomass allocated to above ground wood (branch turnover+stem).

falloccanopy

num. Fraction of biomass allocated to canopy (leaves + reproductive organs + twigs).

Cseedrain

num. Constant used to scale total seed rain per hectare across species.

nbs0

num. Number of seeds produced and dispersed by each mature tree when SEEDTRADEOFF is not defined.

sigma_height

num. Intraspecific variation in tree height (lognormal).

sigma_CR

num. Intraspecific variation in crown radius (lognormal).

sigma_CD

num. Intraspecific variation in crown depth (lognormal).

sigma_P

num. Intraspecific variation in leaf phosphorus (lognormal).

sigma_N

num. Intraspecific variation in leaf nitrogen (lognormal).

sigma_LMA

num. Intraspecific variation in leaf mass per area (lognormal).

sigma_wsg

num. Intraspecific variation in wood specific gravity.

sigma_dbhmax

num. Intraspecific variation in maximum diameter.

corr_CR_height

num. Correlation coefficient between crown radius and tree height.

corr_N_P

num. Correlation coefficient between leaf nitrogen and leaf phosphorus.

corr_N_LMA

num. Correlation coefficient between leaf nitrogen and leaf mass per area

corr_P_LMA

num. Correlation coefficient between leaf phosphorus and leaf mass per area

leafdem_resolution

num. Resolution of leaf demography model.

p_tfsecondary

num. Probability of secondary treefall.

hurt_decay

num. Parameter determining how tree damages are repaired.

crown_gap_fraction

num. Fraction of gaps in the crown.

m

num. Minimal death rate.

m1

num. Slope of death rate m1.

Cair

num. Atmospheric CO2 concentration in micromol/mol.

LL_parameterization

num. Leaf lifespan parameterizations: Reich empirical, Kikuzawa model, and Kikuzawa model with leaf plasticity (0,1,2).

LA_regulation

num. Dynamic LA regulation: off, 1.0, 0.75, or 0.5 (0,1,2,3).

sapwood

num. Sapwood parameterizations: constant thickness (0.04), Fyllas percentage, Fyllas lower limit (0,1,2).

seedsadditional

num. Excess biomass into seeds after maturation (0,1).

NONRANDOM

num. If _NONRANDOM == 1, the seeds for the random number generators will be kept fixed at 1, default for bug fixing (0,1).

GPPcrown

num. This defines an option to compute only GPP from the topmost value of PPFD and GPP, instead of looping within the crown (0,1).

BASICTREEFALL

num. If defined: treefall is a source of tree death (0,1).

SEEDTRADEOFF

num. if defined: the number of seeds produced is determined by NPP allocated to reproduction and seed mass, otherwise the number of seeds is fixed (0,1).

CROWN_MM

num. Michaelis Menten allometry for crowns instead of power law, parameters have to be changed in other input sheets accordingly (0,1).

OUTPUT_extended

num. extended set of ouput files (0,1).

extent_visual

num. extent for visualization output. Unactivated when equal 0.

Value

A data frame of global parameters.

Examples

generate_parameters(nbiter = 12)
#>                   param    value
#> 1                  cols 2.00e+02
#> 2                  rows 2.00e+02
#> 3                HEIGHT 7.00e+01
#> 4          length_dcell 2.50e+01
#> 5                nbiter 1.20e+01
#> 6           iterperyear 1.20e+01
#> 7                    NV 1.00e+00
#> 8                    NH 1.00e+00
#> 9                 nbout 4.00e+00
#> 10                nbspp 4.50e+01
#> 11             SWtoPPFD 2.27e+00
#> 12            p_nonvert 5.00e-02
#> 13               klight 6.30e-01
#> 14                  phi 9.30e-02
#> 15   absorptance_leaves 9.00e-01
#> 16                theta 7.00e-01
#> 17                   g1 3.77e+00
#> 18                   vC 2.10e-02
#> 19                 DBH0 5.00e-03
#> 20                   H0 9.50e-01
#> 21               CR_min 3.00e-01
#> 22                 CR_a 2.13e+00
#> 23                 CR_b 6.30e-01
#> 24                 CD_a 0.00e+00
#> 25                 CD_b 2.00e-01
#> 26                  CD0 3.00e-01
#> 27          shape_crown 7.20e-01
#> 28                 dens 1.00e+00
#> 29           fallocwood 3.50e-01
#> 30         falloccanopy 2.50e-01
#> 31            Cseedrain 5.00e+04
#> 32                 nbs0 1.00e+01
#> 33         sigma_height 0.00e+00
#> 34             sigma_CR 0.00e+00
#> 35             sigma_CD 0.00e+00
#> 36              sigma_P 0.00e+00
#> 37              sigma_N 0.00e+00
#> 38            sigma_LMA 0.00e+00
#> 39            sigma_wsg 0.00e+00
#> 40         sigma_dbhmax 0.00e+00
#> 41       corr_CR_height 0.00e+00
#> 42             corr_N_P 0.00e+00
#> 43           corr_N_LMA 0.00e+00
#> 44           corr_P_LMA 0.00e+00
#> 45   leafdem_resolution 3.00e+01
#> 46        p_tfsecondary 1.00e+00
#> 47           hurt_decay 0.00e+00
#> 48   crown_gap_fraction 1.50e-01
#> 49                    m 1.30e-02
#> 50                   m1 1.30e-02
#> 51                 Cair 4.00e+02
#> 52 _LL_parameterization 1.00e+00
#> 53       _LA_regulation 2.00e+00
#> 54             _sapwood 1.00e+00
#> 55     _seedsadditional 0.00e+00
#> 56           _NONRANDOM 1.00e+00
#> 57            _GPPcrown 0.00e+00
#> 58       _BASICTREEFALL 1.00e+00
#> 59        _SEEDTRADEOFF 0.00e+00
#> 60            _CROWN_MM 0.00e+00
#> 61     _OUTPUT_extended 1.00e+00
#> 62        extent_visual 0.00e+00
#>                                                                                                                                            description
#> 1                                                                                                                                  /* nb of columns */
#> 2                                                                                                                                    /* nb of rows  */
#> 3                                                                                                                  /* vertical extent of simulation */
#> 4                                                                                                                         /* linear size of a dcell */
#> 5                                                                                                                          /* total nb of timesteps */
#> 6                                                                                                                   /* number of iteration per year */
#> 7                                                                                                                /* vertical nb of cells (nb per m) */
#> 8                                                                                                              /* horizontal nb of cells (nb per m) */
#> 9                                                                                                                              /* Number of outputs */
#> 10                                                                                                                             /* Number of species */
#> 11                                                                                       /* convert short wave irradiance to PAR photons (cf. code) */
#> 12                                                                                            /* light incidence param (diff through turbid medium) */
#> 13                                                                                                  /* light attenuation in the canopy Beer-Lambert */
#> 14                                                                                                /*  quantum yield (in micromol C/micromol photon) */
#> 15                                                                                                              /* absorptance of individual leaves */
#> 16                                                                                                               /* parameter of the Farquhar model */
#> 17                                                                                     /* parameter g1 of Medlyn et al s stomatal conductance model */
#> 18                                                                                                                /* variance of the flexion moment */
#> 19                                                                                                                               /* initial dbh (m) */
#> 20                                                                                                                            /* initial height (m) */
#> 21                                                                                                                   /* minimum crown radius (in m) */
#> 22                                                                                           /* CR log intercept or Michaelis Menten initial growth */
#> 23                                                                                                /* CR log slope or Michaelis Menten asymptotic CR */
#> 24                                                                                                                 /* CD intercept (absolute value) */
#> 25                                                                                                         /* CD slope (as fraction of tree height) */
#> 26                                                                                                                     /* initial crown depth(in m) */
#> 27                                                                                                                         /* crown shape parameter */
#> 28                                                                                                                /* initial leaf density (m^2/m^2) */
#> 29                                                                     /* fraction of biomass allocated to above ground wood (branch turnover+stem) */
#> 30                                                                /* fraction of biomass allocated to canopy (leaves + reproductive organs + twigs) */
#> 31                                                       /* constant used to scale total seed rain per hectare across species (in next computation) */
#> 32                                                       /* nb of seeds produced and dispersed by each mature tree when SEEDTRADEOFF is not defined */
#> 33                                                                                            /* intraspecific variation in tree height (lognormal) */
#> 34                                                                                           /* intraspecific variation in crown radius (lognormal) */
#> 35                                                                                            /* intraspecific variation in crown depth (lognormal) */
#> 36                                                                                        /* intraspecific variation in leaf phosphorus (lognormal) */
#> 37                                                                                          /* intraspecific variation in leaf nitrogen (lognormal) */
#> 38                                                                                                    /* intraspecific variation in LMA (lognormal) */
#> 39                                                                                              /* intraspecific variation in wood specific gravity */
#> 40                                                                                                   /* intraspecific variation in maximum diameter */
#> 41                                                                                  /* correlation coefficient between crown radius and tree height */
#> 42                                                                             /* correlation coefficient between leaf nitrogen and leaf phosphorus */
#> 43                                                                                         /* correlation coefficient between leaf nitrogen and LMA */
#> 44                                                                                       /* correlation coefficient between leaf phosphorus and LMA */
#> 45                                                                                                           /* resolution of leaf demography model */
#> 46                                                                                                             /* probability of secondary treefall */
#> 47                                                                                          /*  parameter determining how tree damages are repaired */
#> 48                                                                                                                 /* fraction of gaps in the crown */
#> 49                                                                                                                            /* minimal death rate */
#> 50                                                                                                                      /* m1 (slope of death rate) */
#> 51                                                                                                 /* atmospheric CO2 concentration in micromol/mol */
#> 52                                        /* LL parameterizations: Reich empirical, Kikuzawa model, and Kikuzawa model with leaf plasticity (0,1,2) */
#> 53                                                                                       /* dynamic LA regulation: off, 1.0, 0.75, or 0.5 (0,1,2,3) */
#> 54                                           /* sapwood parameterizations: constant thickness (0.04), Fyllas percentage, Fyllas lower limit (0,1,2) */
#> 55                                                                                              /* excess biomass into seeds after maturation (0,1) */
#> 56                                /* If _NONRANDOM == 1, the seeds for the random number generators will be kept fixed at 1, default for bug fixing */
#> 57                       /* This defines an option to compute only GPP from the topmost value of PPFD and GPP, instead of looping within the crown. */
#> 58                                                                                                /* if defined: treefall is a source of tree death */
#> 59 /* if defined: the number of seeds produced is determined by NPP allocated to reproduction and seed mass, otherwise the number of seeds is fixed */
#> 60                   /* Michaelis Menten allometry for crowns instead of power law, parameters have to be changed in other input sheets accordingly */
#> 61                                                                                                                   /* extended set of ouput files */
#> 62                                                                                                                /* extent for visualization output *